Payments

+++ src/controllers/contracts/individualController.js
@@ -1,6 +1,7 @@
 const { SUPER } = require('../../constants/roles');
-const { ContratoIndividual, ContratoGeneral, Institucion, Pasajero, Parametro, Cuota } = require('../../database/models');
+const { ContratoIndividual, ContratoGeneral, Institucion, Pasajero, Parametro, Cuota, Movimiento } = require('../../database/models');
 const { validationResult } = require('express-validator');
+const { Op, where } = require('sequelize');
 
 module.exports = {
   get: async (req, res) => {
@@ -24,7 +25,7 @@ module.exports = {
       });
       if (req.user.rol.name !== SUPER) {
         const mapped = individualContracts?.map((result) => result.dataValues);
-        individualContracts = mapped?.filter((el) => el.estado === 'vigente');
+        individualContracts = mapped?.filter((el) => el.estado === 'vigente' || el.estado === 'pagado');
       }
       res.status(200).json({
         status: 'success',
@@ -39,9 +40,54 @@ module.exports = {
       });
     }
   },
+  getCodes: async (req, res) => {
+    try {
+      const { id } = req.query;
+      let individualContracts = await ContratoIndividual.findAll({
+        include: [
+          {
+            model: Pasajero,
+            as: 'pasajero'
+          },
+          {
+            model: ContratoGeneral,
+            as: 'contrato_general',
+            include: [
+              {
+                model: Institucion,
+                as: 'institucion'
+              }
+            ]
+          }
+        ],
+        order: [['id', 'DESC']]
+      });
+      if (id) {
+        const mapped = individualContracts.filter((el) => el.id == id);
+        individualContracts = mapped;
+      }
+      const data = individualContracts
+        .filter((el) => el.estado === 'vigente' /* || el.estado === 'pagado' */)
+        .map((el) => ({
+          label: `${el.pasajero.documento} - ${el.pasajero.apellido}, ${el.pasajero.nombre} # ${el.contrato_general.institucion.nombre} - Grado: ${el.contrato_general.grado} - División: ${el.contrato_general.division} - Turno:  ${el.contrato_general.turno} # (${el.contrato_general.descripcion})`,
+          id: el.id
+        }));
+      res.status(200).json({
+        status: 'success',
+        count: individualContracts.length,
+        data
+      });
+    } catch (error) {
+      res.status(409).json({
+        msg: 'Ha ocurrido un error al intentar recuperar los responsables',
+        error,
+        status: 'error'
+      });
+    }
+  },
   getByQuery: async (req, res) => {
     try {
-      const { code } = req.query;
+      const { code, document } = req.query;
       let individualContracts;
       if (code) {
         individualContracts = await ContratoIndividual.findAll({
@@ -49,7 +95,35 @@ module.exports = {
           include: [
             {
               model: ContratoGeneral,
-              as: 'contrato_general'
+              as: 'contrato_general',
+              include: {
+                model: Institucion,
+                as: 'institucion'
+              }
+            },
+            {
+              model: Pasajero,
+              as: 'pasajero'
+            }
+          ],
+          order: [['id', 'DESC']]
+        });
+      }
+      if (document) {
+        individualContracts = await ContratoIndividual.findAll({
+          where: {
+            cod_contrato: {
+              [Op.like]: `%${document}%`
+            }
+          },
+          include: [
+            {
+              model: ContratoGeneral,
+              as: 'contrato_general',
+              include: {
+                model: Institucion,
+                as: 'institucion'
+              }
             },
             {
               model: Pasajero,
@@ -61,7 +135,7 @@ module.exports = {
       }
       if (req.user.rol.name !== SUPER) {
         const mapped = individualContracts?.map((result) => result.dataValues);
-        individualContracts = mapped?.filter((el) => el.estado === 'vigente');
+        individualContracts = mapped?.filter((el) => el.estado === 'vigente' || el.estado === 'pagado');
       }
       return res.status(200).json({
         status: 'success',
@@ -91,7 +165,7 @@ module.exports = {
         where: {
           id_contrato_individual: id
         },
-        attributes: ['valor_primer_vencimiento', 'valor_segundo_vencimiento', 'estado', 'numero']
+        attributes: ['id', 'valor_primer_vencimiento', 'valor_segundo_vencimiento', 'estado', 'numero']
       });
 
       const parametros = await Parametro.findOne({
@@ -138,6 +212,7 @@ module.exports = {
         cuotasRecalculadas,
         cuotasPagadas,
         senia,
+        newContractValue: nuevo_valor,
         data: {
           cuotasPendientes,
           diferencia_precio,
@@ -152,8 +227,26 @@ module.exports = {
       });
     }
   },
+  newShares: async (req, res) => {
+    try {
+      let { id } = req.params;
+      let { shares, newContractValue } = req.body;
+      await ContratoIndividual.update({ valor_contrato: newContractValue }, { where: { id } });
+      await Promise.all(shares.map((el) => Cuota.update({ ...el }, { where: { id: el.id } })));
+      res.status(200).json({
+        status: 'success',
+        msg: 'Nuevas cuotas actualizadas. Nuevo valor del Contrato Individual:' + '$ ' + newContractValue,
+        shares
+      });
+    } catch (error) {
+      res.status(409).json({
+        status: 'error',
+        msg: 'Ha ocurrido un error al intentar recalcular las cuotas de un contrato individual',
+        error
+      });
+    }
+  },
   getById: (req, res) => {
-    console.log('AAAAAAAAAAAAAAAAAAAAAAAAAAAAACAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA');
     res.status(200).json({
       status: 'success',
       msg: 'Contrato individual encontrado',
@@ -179,6 +272,15 @@ module.exports = {
       try {
         const { id_contrato_general, id_pasajero, valor, cuotas, codigo_contrato_individual } = req.body;
 
+        const individualContractExist = await ContratoIndividual.findOne({ where: { cod_contrato: codigo_contrato_individual } });
+
+        if (individualContractExist) {
+          return res.status(400).json({
+            status: 'success',
+            msg: 'Ya existe un Contrato Individual para este pasajero'
+          });
+        }
+
         const individualContract = await ContratoIndividual.create({
           id_contrato_general: id_contrato_general,
           id_pasajero: id_pasajero,
@@ -199,9 +301,10 @@ module.exports = {
 
         res.status(200).json({
           status: 'success',
-          msg: 'Contrato individual creado con éxito',
+          msg: 'Contrato individual creado con éxito. Redireccionando a pagos',
           data: {
-            codigo: individualContract.cod_contrato
+            codigo: individualContract.cod_contrato,
+            id: individualContract.id
           }
         });
       } catch (error) {
@@ -226,6 +329,112 @@ module.exports = {
       try {
         const { estado } = req.body;
         const { id } = req.params;
+        const { individualContract } = req;
+        if (estado === 'cancelado') {
+          if (individualContract.estado === 'terminado') {
+            // Eliminicación de todas las cuotas
+            await Cuota.destroy({ where: { id_contrato_individual: id } });
+
+            // Acualización del estado del contrato individual
+            await ContratoIndividual.update({ estado }, { where: { id } });
+
+            return res.status(200).json({
+              status: 'success',
+              msg: `Contrato individual editado con éxito. Se eliminaron todas las cuotas asociadas`
+            });
+          }
+
+          // Obtención de todas las cuotas asociadas al Contrato Individual que será borrado
+          const installments = await Cuota.findAll({
+            where: { id_contrato_individual: id, numero: { [Op.not]: 0 }, estado: 'pagada' },
+            attributes: ['id', 'valor_primer_vencimiento']
+          });
+
+          // Calculo del total a devolver
+          const total_return = installments.reduce((acc, el) => acc + Number(el.valor_primer_vencimiento), 0) * -1;
+
+          // Inserción del movimiento(egreso)
+          await Movimiento.create({
+            importe: total_return,
+            tipo: 'egreso',
+            forma_pago: 'egreso',
+            info: `Devolución de cuotas por eliminación de Contrato Individual ${individualContract.cod_contrato}`
+          });
+
+          // Eliminicación de todas las cuotas
+          await Cuota.destroy({ where: { id_contrato_individual: id } });
+
+          // Acualización del estado del contrato individual
+          await ContratoIndividual.update({ estado }, { where: { id } });
+
+          // Actualización del lo asientos ocupados del contrato general
+          const generalContract = await ContratoGeneral.findOne({ where: { id: individualContract.id_contrato_general } });
+          await ContratoGeneral.update(
+            { asientos_ocupados: Number(generalContract.asientos_ocupados) - 1 },
+            { where: { id: individualContract.id_contrato_general } }
+          );
+
+          return res.status(200).json({
+            status: 'success',
+            cuotas_devueltas: installments,
+            total_cuotas_devultas: total_return,
+            msg: `Contrato individual editado con éxito. Se liberó un asiento del Contrato general. Se devolvieron $ ${
+              total_return * -1
+            } en concepto de cuotas.  Se eliminaron todas las cuotas asociadas`
+          });
+        }
+
+        if (estado === 'pagado') {
+          // Obtención de todas las cuotas asociadas al Contrato Individual que será borrado
+          const installments = await Cuota.findAll({
+            where: { id_contrato_individual: id, numero: { [Op.not]: 0 }, estado: 'pendiente' },
+            attributes: ['id', 'valor_primer_vencimiento']
+          });
+
+          // Calculo del total a ingresar
+          const total_charge = installments.reduce((acc, el) => acc + Number(el.valor_primer_vencimiento), 0);
+
+          // Inserción del movimiento(ingreso)
+          let msg = `Contrato individual editado con éxito.`;
+          if (total_charge !== 0) {
+            await Movimiento.create({
+              importe: total_charge,
+              tipo: 'ingreso',
+              forma_pago: 'efectivo',
+              info: `Pago de todas las cuotas pendientes del Contrato Individual ${individualContract.cod_contrato}`
+            });
+
+            msg = `Contrato individual editado con éxito. Ingresaron $ ${total_charge} en concepto de cuotas`;
+          }
+
+          // Actualización del estado de todas las cuotas
+          await Promise.all(
+            installments.map(async (el) => await Cuota.update({ estado: 'pagada' }, { where: { id_contrato_individual: id } }))
+          );
+
+          // Acualización del estado del contrato individual
+          await ContratoIndividual.update({ estado, pagos: Number(individualContract.pagos) + Number(total_charge) }, { where: { id } });
+
+          return res.status(200).json({
+            status: 'success',
+            cuotas_pagadas: installments,
+            total_cuotas_pagadas: total_charge,
+            msg
+          });
+        }
+
+        if (estado === 'terminado') {
+          const individualContract = await ContratoIndividual.findByPk(id);
+          const saldado = Number(individualContract.valor_contrato) < Number(individualContract.pagos);
+
+          if (!saldado) {
+            return res.status(400).json({
+              status: 'error',
+              msg: 'Existen cuotas impagas. Imposible terminar contrato'
+            });
+          }
+        }
+
         await ContratoIndividual.update({ estado }, { where: { id } });
         res.status(200).json({
           status: 'success',
@@ -250,14 +459,41 @@ module.exports = {
   delete: async (req, res) => {
     try {
       const { id } = req.params;
-      await ContratoIndividual.destroy({ where: { id } });
-      res.status(200).json({
-        status: 'success',
-        msg: 'Contrato general borrado con éxito'
-      });
+
+      const { individualContract } = req;
+
+      if (individualContract.estado === 'terminado') {
+        // Eliminicación de todas las cuotas
+        await Cuota.destroy({ where: { id_contrato_individual: id } });
+
+        // Eliminación del contrato general
+        await ContratoIndividual.destroy({ where: { id } });
+
+        return res.status(200).json({
+          status: 'success',
+          msg: 'Contrato individual eliminado con éxito. Se eliminaron todas las cuotas asociadas'
+        });
+      }
+
+      if (individualContract.estado === 'vigente') {
+        return res.status(400).json({
+          status: 'success',
+          msg: 'No es posible eliminar un contrato vigente'
+        });
+      }
+
+      if (individualContract.estado === 'cancelado') {
+        // Eliminación del contrato general
+        await ContratoIndividual.destroy({ where: { id } });
+
+        return res.status(200).json({
+          status: 'success',
+          msg: 'Contrato individual eliminado con éxito. Se eliminaron todas las cuotas asociadas'
+        });
+      }
     } catch (error) {
       res.status(409).json({
-        msg: 'Ha ocurrido un error al intentar borrar el contrato general',
+        msg: 'Ha ocurrido un error al intentar borrar el contrato individual',
         error,
         status: 'error'
       });
